# src/main/resources/swagger-info.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:8080",
	"basePath": "/",
	"tags": [
		{
			"name": "basic-error-controller",
			"description": "Basic Error Controller"
		},
		{
			"name": "task-controller",
			"description": "Task Controller"
		},
		{
			"name": "user-controller",
			"description": "User Controller"
		}
	],
	"paths": {
		"/login": {
			"post": {
				"tags": [
					"user-controller"
				],
				"summary": "createAuthenticationToken",
				"operationId": "createAuthenticationTokenUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "userModel",
						"description": "userModel",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserModel"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/apiejemplo/deletetask": {
			"delete": {
				"tags": [
					"task-controller"
				],
				"summary": "deleteTask",
				"operationId": "deleteTaskUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "taskModel",
						"description": "taskModel",
						"required": true,
						"schema": {
							"$ref": "#/definitions/TaskModel"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"deprecated": false
			}
		},
		"/apiejemplo/finishtask": {
			"put": {
				"tags": [
					"task-controller"
				],
				"summary": "finisTask",
				"operationId": "finisTaskUsingPUT",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "taskModel",
						"description": "taskModel",
						"required": true,
						"schema": {
							"$ref": "#/definitions/TaskModel"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/apiejemplo/savetask": {
			"post": {
				"tags": [
					"task-controller"
				],
				"summary": "saveTask",
				"operationId": "saveTaskUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "listTask",
						"description": "listTask",
						"required": true,
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/TaskModel"
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
		"/apiejemplo/showtask": {
			"get": {
				"tags": [
					"task-controller"
				],
				"summary": "showTask",
				"operationId": "showTaskUsingGET",
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		}
	},
	"definitions": {
		"TaskModel": {
			"type": "object",
			"properties": {
				"description": {
					"type": "string"
				},
				"ended": {
					"type": "boolean"
				},
				"title": {
					"type": "string"
				},
				"userId": {
					"type": "string"
				}
			},
			"title": "TaskModel"
		},
		"UserModel": {
			"type": "object",
			"properties": {
				"password": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			},
			"title": "UserModel"
		}
	}
}